+++ cropclassification/util/openeo_util.py
@@ -1,15 +1,21 @@
-from collections import defaultdict
-from datetime import datetime
 import json
 import logging
-from pathlib import Path
 import pprint
 import time
+from collections import defaultdict
+from datetime import datetime
+from pathlib import Path
 from typing import Any, Optional
 
+import numpy as np
 import openeo
+import openeo.rest.datacube
 import openeo.rest.job
 import pyproj
+from openeo.api.process import Parameter
+from openeo.processes import if_, is_nan
+from scipy.signal.windows import gaussian  # type: ignore  # noqa: PGH003
+from shapely.geometry import Polygon
 
 from . import raster_util
 from .io_util import output_exists
@@ -34,6 +40,7 @@ def get_images(
       - roi_bounds (Tuple): bounds of the image to get as Tuple[minx, miny, maxx, maxy].
       - roi_crs (int): crs the roi bounds.
       - collection (str): openeo collection to get the image from.
+      - satellite (str): the satellite to get the image from.
       - start_date (datetime): start date for the images in the collection to use for
         this image.
       - end_date (datetime): exclusive end date for the images in the collection to use
@@ -64,7 +71,6 @@ def get_images(
         ValueError: _description_
         Exception: _description_
         RuntimeError: _description_
-        RuntimeError: _description_
     """
     # Do some checks on already existing images
     images_to_get_todo = {}
@@ -89,6 +95,8 @@ def get_images(
     # More info on the configuration file format can be found here:
     # https://open-eo.github.io/openeo-python-client/configuration.html#configuration-files  # noqa: E501
     conn = openeo.connect()
+    # conn.authenticate_oidc_refresh_token()
+
     logger.info(conn.describe_account())
 
     images_to_get_dict = {image["path"].as_posix(): image for image in images_to_get}
@@ -145,6 +153,7 @@ def get_images(
         create_mosaic_job(
             conn=conn,
             collection=image_to_get["collection"],
+            satellite=image_to_get["satellite"],
             spatial_extent=roi_extent,
             start_date=image_to_get["start_date"],
             end_date=image_to_get["end_date"],
@@ -174,6 +183,7 @@ def get_images(
 def create_mosaic_job(
     conn: openeo.Connection,
     collection: str,
+    satellite: str,
     spatial_extent,
     start_date: datetime,
     end_date: datetime,
@@ -217,46 +227,65 @@ def create_mosaic_job(
             "used together"
         )
 
-    # Load cube of relevant images
-    # You can look which layers are available here:
-    # https://openeo.cloud/data-collections/
-    cube = conn.load_collection(
-        collection,
-        spatial_extent=spatial_extent,
-        temporal_extent=period,
-        bands=bands_to_load,
-        max_cloud_cover=max_cloud_cover,
-    )
+    if time_reducer == "best":
+        if cloud_filter_band is not None or cloud_filter_band_dilated is not None:
+            raise ValueError(
+                "process_option cloud_filter_band and cloud_filter_band_dilated "
+                "are not being used with time_reducer = best"
+            )
 
-    # Use mask_scl_dilation for "aggressive" cloud mask based on SCL
-    if cloud_filter_band_dilated is not None:
-        cube = cube.process(
-            "mask_scl_dilation",
-            data=cube,
-            scl_band_name=cloud_filter_band,
+        if satellite == "s1":
+            raise ValueError("only Sentinel 2 can be used with time_reducer = best")
+
+        cube = best_available_pixel(
+            conn=conn,
+            collection=collection,
+            spatial_extent=spatial_extent,
+            temporal_extent=period,
+            bands=bands,
+            max_cloud_cover=max_cloud_cover,
         )
-    elif cloud_filter_band is not None:
-        # Select the scl band from the data cube
-        scl_band = cube.band(cloud_filter_band)
-        # Build mask to mask out everything but classes
-        #   - 4: vegetation
-        #   - 5: non vegetated
-        #   - 6: water
-        mask = ~((scl_band == 4) | (scl_band == 5) | (scl_band == 6))
+    else:
+        # Load cube of relevant images
+        # You can look which layers are available here:
+        # https://openeo.cloud/data-collections/
+        cube = conn.load_collection(
+            collection,
+            spatial_extent=spatial_extent,
+            temporal_extent=period,
+            bands=bands_to_load,
+            max_cloud_cover=max_cloud_cover,
+        )
 
-        # The mask needs to have the same "ground sample distance" as the cube it is
-        # applied to.
-        mask = mask.resample_cube_spatial(cube)
+        # Use mask_scl_dilation for "aggressive" cloud mask based on SCL
+        if cloud_filter_band_dilated is not None:
+            cube = cube.process(
+                "mask_scl_dilation",
+                data=cube,
+                scl_band_name=cloud_filter_band,
+            )
+        elif cloud_filter_band is not None:
+            # Select the scl band from the data cube
+            scl_band = cube.band(cloud_filter_band)
+            # Build mask to mask out everything but classes
+            #   - 4: vegetation
+            #   - 5: non vegetated
+            #   - 6: water
+            mask = ~((scl_band == 4) | (scl_band == 5) | (scl_band == 6))
 
-        # Apply the mask
-        cube = cube.mask(mask)
+            # The mask needs to have the same "ground sample distance" as the cube it is
+            # applied to.
+            mask = mask.resample_cube_spatial(cube)
 
-    cube = cube.filter_bands(bands=bands)
-    cube = cube.reduce_dimension(dimension="t", reducer=time_reducer)
+            # Apply the mask
+            cube = cube.mask(mask)
 
-    # The NDVI collection needs to be recalculated
-    if collection == "TERRASCOPE_S2_NDVI_V2":
-        cube = cube.apply(lambda x: 0.004 * x - 0.08)
+        cube = cube.filter_bands(bands=bands)
+        cube = cube.reduce_dimension(dimension="t", reducer=time_reducer)
+
+        # The NDVI collection needs to be recalculated
+        if collection == "TERRASCOPE_S2_NDVI_V2":
+            cube = cube.apply(lambda x: 0.004 * x - 0.08)
 
     return cube.create_job(
         out_format="GTiff",
@@ -310,8 +339,9 @@ def get_job_results(
                         output_tmp_path.unlink()
 
                     asset = batch_job.get_results().get_asset()
-                    import requests
                     import shutil
+
+                    import requests
 
                     def download_file(url, target):
                         with requests.get(url, stream=True) as r:
@@ -377,3 +407,188 @@ def postprocess_image(path: Path, band_descriptions: Optional[list[str]]):
         raster_util.set_band_descriptions(path, band_descriptions)
     else:
         logger.warning(f"no band_descriptions specified for {path}")
+
+
+def best_available_pixel(
+    conn: openeo.Connection,
+    collection: str,
+    spatial_extent,
+    temporal_extent: list[str],
+    bands: list[str],
+    max_cloud_cover: Optional[float],
+):
+    """
+    Create image mosaic using the Best Available Pixel (BAP) method.
+
+    Link to the article:
+    https://documentation.dataspace.copernicus.eu/APIs/openEO/openeo-community-examples/python/RankComposites/bap_composite.html
+
+    Args:
+        conn (openeo.Connection): openeo connection.
+        collection (str): openeo collection to get the image from.
+        spatial_extent (_type_): bounds of the image
+        temporal_extent (list[str]): period between start and end date
+        bands (list[str]): the bands to get.
+        max_cloud_cover (Optional[float]): the maximum cloud cover images
+        should have to be used.
+
+    Returns:
+        DataCube: _description_
+    """
+    spatial_resolution = 20
+
+    scl = conn.load_collection(
+        collection,
+        spatial_extent=spatial_extent,
+        temporal_extent=temporal_extent,
+        bands=["SCL"],
+        max_cloud_cover=max_cloud_cover,
+    ).resample_spatial(spatial_resolution)
+    scl = scl.apply(lambda x: if_(is_nan(x), 0, x))
+
+    classification = scl.band("SCL")
+    binary = (
+        (classification == 3)
+        | (classification == 8)
+        | (classification == 9)
+        | (classification == 10)
+    )
+    binary = binary.add_dimension(name="bands", label="score", type="bands")
+
+    # Coverage Score
+    # --------------
+    # First of all, the coverage score is calculated, which is equal to 1 minus
+    # the percentage of pixels in the spatial_extent that are classified as cloud in SCL
+
+    # A polygon is created that is slighty larger than the original spatial_extent
+
+    # The cloud coverage percentage is calculated by taking the average
+    # of the cloud binary (defined above), by using the aggregate_spatial process
+
+    # Since the aggregate_spatial process results in vectordata,
+    # the data is rasterized again to the original dimensions,
+    # using the (experimental) vector_to_raster process
+
+    # The coverage score is then equal to 1 - the cloud coverage percentage.
+    # I.e. a smaller cloud coverage percentage indicates a larger score
+
+    min_lon, max_lon, min_lat, max_lat = list(spatial_extent.values())
+    polygon = Polygon(
+        [
+            (min_lon, min_lat),
+            (min_lon, max_lat),
+            (max_lon, max_lat),
+            (max_lon, min_lat),
+            (min_lon, min_lat),
+        ]
+    )
+
+    cloud_coverage = (1 - binary).aggregate_spatial(geometries=polygon, reducer="mean")
+
+    # rasterize the vectorcube so result has the same spatial dimensions as scl cube.
+    coverage_score = cloud_coverage.vector_to_raster(scl)
+    coverage_score = coverage_score.rename_labels("bands", ["score"])
+
+    # Data Score
+    # ----------
+    # Both the date and distance-to-cloud score are calculated via a UDF.
+    label = Parameter("label")
+
+    def day_of_month_calc(input):
+        day = lambda x: 15 + x.process(
+            "date_difference",
+            date1=x.process(
+                "date_replace_component", date=label, value=15, component="day"
+            ),
+            date2=label,
+            unit="day",
+        )
+        return input.array_apply(day)
+
+    day_of_month = scl.apply_neighborhood(
+        day_of_month_calc,
+        size=[
+            {"dimension": "x", "unit": "px", "value": 1},
+            {"dimension": "y", "unit": "px", "value": 1},
+            {"dimension": "t", "value": "month"},
+        ],
+        overlap=[],
+    )
+
+    def get_date_score(day) -> openeo.rest.datacube.DataCube:
+        return (
+            day.subtract(15)
+            .multiply(0.2)
+            .multiply(day.subtract(15).multiply(0.2))
+            .multiply(-0.5)
+            .exp()
+            .multiply(0.07978845)
+        )  # Until 'power' and 'divide' are fixed, use this workaround
+
+    date_score = (1.0 * day_of_month).apply(get_date_score)
+    date_score = date_score.rename_labels("bands", ["score"])
+
+    # Distance to Cloud Score
+    # Finally, the Distance to Cloud Score is calculated by applying a Gaussian kernel
+    # to the binary. The kernel size, defined below, indicates the maximum distance
+    # from a cloud, above which a pixel is automatically assigned score 1.
+    # In case of a kernel size of 151, the maximum distance is 150 pixels.
+    # The kernel size is 151 for a resolution of 20m and will be rescaled accordingly.
+
+    kernel_size = int(150 * 20 / spatial_resolution) + 1
+    gaussian_1d = gaussian(M=kernel_size, std=1)
+    gaussian_kernel = np.outer(gaussian_1d, gaussian_1d)
+    gaussian_kernel /= gaussian_kernel.sum()
+
+    dtc_score = 1 - binary.apply_kernel(gaussian_kernel)
+
+    # Aggregating Scores
+    # Then, the scores are aggregated together, by using a weighted average.
+    # As for now the coverage score is excluded.
+    # To make sure that pixels which contain no data (SCL score 0) are not selected,
+    # they are explicitly masked out of the final score. B01 score is excluded.
+    weights = [1, 0.8, 0.5]
+    score = (
+        weights[0] * dtc_score + weights[1] * date_score + weights[2] * coverage_score
+    ) / sum(weights)
+    score = score.mask(classification == 0)
+
+    # Masking and Compositing
+    # Next, a mask is created. This serves to mask every pixel,
+    # except the one with the highest score, for each month.
+    def max_score_selection(score):
+        max_score = score.max()
+        return score.array_apply(lambda x: x != max_score)
+
+    rank_mask = score.apply_neighborhood(
+        max_score_selection,
+        size=[
+            {"dimension": "x", "unit": "px", "value": 1},
+            {"dimension": "y", "unit": "px", "value": 1},
+            {"dimension": "t", "value": "month"},
+        ],
+        overlap=[],
+    )
+
+    rank_mask = rank_mask.band("score")
+
+    # Next, some bands of interest from Sentinel-2 are loaded. They are then masked by
+    # the BAP mask constructed above.
+    # Then they are aggregated per period, to obtain a composite image per period.
+    # By using the "first" process as an aggregator, the situation where there are
+    # more than one images selected in the period for a pixel, is handled.
+    rgb_bands = conn.load_collection(
+        collection,
+        spatial_extent=spatial_extent,
+        temporal_extent=temporal_extent,
+        bands=bands,
+        max_cloud_cover=max_cloud_cover,
+    )
+
+    composite = (
+        rgb_bands.mask(rank_mask)
+        .mask(binary)
+        .aggregate_temporal_period("month", "first")
+    )
+
+    return composite