+++ tests/test_openeo_util.py
@@ -1,5 +1,6 @@
-from datetime import datetime
+import os
 import shutil
+from datetime import datetime
 
 import pytest
 import rasterio
@@ -60,6 +61,7 @@ def test_get_images_s1_edge_first(tmp_path):
     test_filename = "s1-on-border-first.tif"
     output_path = tmp_path / test_filename
     shutil.copy(test_helper.testdata_dir / test_filename, output_path)
+
     images_to_get = [
         {
             "path": output_path,
@@ -80,7 +82,8 @@ def test_get_images_s1_edge_first(tmp_path):
     assert output_path.exists()
 
 
-def test_get_images_s2(tmp_path):
+@pytest.mark.parametrize("time_reducer", ["mean", "best"])
+def test_get_images_s2(tmp_path, time_reducer):
     """
     Test creating an s2 image.
 
@@ -88,20 +91,27 @@ def test_get_images_s2(tmp_path):
     avoid really calling the openeo API.
     """
     # Prepare test data
-    output_path = tmp_path / SampleData.image_s2_path.name
-    shutil.copy(SampleData.image_s2_path, output_path)
-    bands = ["B02", "B03", "B04", "B08", "B11", "B12"]
+    if time_reducer == "best":
+        output_path = tmp_path / SampleData.image_s2_best_path.name
+        shutil.copy(SampleData.image_s2_best_path, output_path)
+    elif time_reducer == "mean":
+        output_path = tmp_path / SampleData.image_s2_mean_path.name
+        shutil.copy(SampleData.image_s2_mean_path, output_path)
+    else:
+        raise ValueError(f"Unknown time_reducer {time_reducer}")
 
+    bands = ["B02", "B03", "B04", "B08", "B11", "B12"]
     images_to_get = [
         {
             "path": output_path,
             "roi_bounds": SampleData.roi_bounds,
             "roi_crs": SampleData.roi_crs,
             "collection": "TERRASCOPE_S2_TOC_V2",
+            "satellite": "s2",
             "start_date": SampleData.start_date,
             "end_date": SampleData.end_date,
             "bands": bands,
-            "time_reducer": "mean",
+            "time_reducer": time_reducer,
             "max_cloud_cover": 80,
             "process_options": {},
             "job_options": {},
@@ -115,3 +125,45 @@ def test_get_images_s2(tmp_path):
         assert file.count == len(bands)
         # All bands should have a description
         assert all(file.descriptions)
+
+
+@pytest.mark.skipif(
+    "GITHUB_ACTIONS" in os.environ,
+    reason="Don't run on CI: not possible to connect to openeo.",
+)
+@pytest.mark.parametrize(
+    "expected_error, satellite, time_reducer",
+    [("only Sentinel 2 can be used with time_reducer = best", "s1", "best")],
+)
+def test_get_images_invalid_params(
+    tmp_path, expected_error: str, satellite: str, time_reducer: str
+):
+    """
+    Test creating an s2 image.
+
+    Remark: the default way to run is with the result pre-copied and force=False, to
+    avoid really calling the openeo API.
+    """
+    # Prepare test data
+    output_path = tmp_path / SampleData.image_s2_best_path.name
+    bands = ["B02", "B03", "B04", "B08", "B11", "B12"]
+    images_to_get = [
+        {
+            "path": output_path,
+            "roi_bounds": SampleData.roi_bounds,
+            "roi_crs": SampleData.roi_crs,
+            "collection": "TERRASCOPE_S2_TOC_V2",
+            "satellite": satellite,
+            "start_date": SampleData.start_date,
+            "end_date": SampleData.end_date,
+            "bands": bands,
+            "time_reducer": time_reducer,
+            "max_cloud_cover": 80,
+            "process_options": {},
+            "job_options": {},
+        }
+    ]
+
+    # Test
+    with pytest.raises(ValueError, match=expected_error):
+        openeo_util.get_images(images_to_get=images_to_get)