+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Add whether a parcel has been used for training to output (#107)
 - Run `bulk_zonal_stats` in low priority worker processes (#81)
 - Use ruff instead of black and flake for formatting and linting (#57, #64, #65, #67)
+- Add method "best available pixel" on openeo for S2 (#70)
 
 ### Bugs fixed